requirements.txt
aiogram>=3.7,<4.0
fastapi>=0.115
uvicorn[standard]>=0.30
python-dotenv>=1.0.1

